fix: change checkDateLater function location to util
